Extract Rails::Command::Spellchecker
