+++ railties/lib/rails/command.rb
@@ -11,6 +11,7 @@ module Rails
   module Command
     extend ActiveSupport::Autoload
 
+    autoload :Spellchecker
     autoload :Behavior
     autoload :Base
